Integrate exercise codes and remove duplicate
